Set up RAA demo
Tail = 1.005 and wtd avg link ratios selected throughout. Kirsten, the
LINEST file is the version of RAA before LINEST calculations were
removed. It is envisioned that you will put those calculations into the
RAA workbook, and select the simple average for the first factor, and
push to github for my review.